feature-policy/reporting/wakelock-report-only.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

const check_report_format = ([reports, observer]) => {
  let report = reports[0];
  assert_equals(report.type, "feature-policy-violation");
  assert_equals(report.body.featureId, "wake-lock");
  assert_equals(report.body.disposition, "report");
};

promise_test(async t => {
  const report = new Promise(resolve => {
    new ReportingObserver((reports, observer) => resolve([reports, observer]),
                          {types: ['feature-policy-violation']}).observe();
  });
  try {
    await WakeLock.request("screen");
    check_report_format(await report);
  } catch (err) {
    // In case the show call was rejected, the report should be generated
    // anyway. Wait for it and check the format before failing this test.
    check_report_format(await report);
    throw err;
  }
}, "WakeLock report only mode");
</script>
